% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{add_colors_to_nodes}
\alias{add_colors_to_nodes}
\title{Add color palettes}
\usage{
add_colors_to_nodes(nodes_df)
}
\arguments{
\item{nodes_df}{Data frame of nodes with 'value' and 'source' columns.}
}
\value{
nodes_df with colored nodes based on their values.
}
\description{
Add color palettes
}
