% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{add_compound_names}
\alias{add_compound_names}
\title{Add compound names to compound nodes in the nodes data frame.}
\usage{
add_compound_names(nodes_df, bfc)
}
\arguments{
\item{nodes_df}{Data frame of nodes with a column 'type' indicating node type}
}
\value{
Updated nodes data frame with compound names added to compound nodes.
}
\description{
Add compound names to compound nodes in the nodes data frame.
}
