% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{add_gene_names}
\alias{add_gene_names}
\title{Parse KEGG KGML files to extract relations (edges) data frame.}
\usage{
add_gene_names(nodes_df)
}
\arguments{
\item{file}{Path to the KGML XML file.}
}
\value{
A tibble
}
\description{
Parse KEGG KGML files to extract relations (edges) data frame.
}
