% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{add_results_nodes}
\alias{add_results_nodes}
\title{Add results from combined results data frame to nodes data frame}
\usage{
add_results_nodes(nodes_df, results_combined)
}
\arguments{
\item{nodes_df}{Data frame of nodes with a column 'KEGG'}

\item{results_combined}{Data frame with combined results containing columns: KEGG, value, source}
}
\value{
Updated nodes data frame with added columns: value, color, source, text
}
\description{
Add results from combined results data frame to nodes data frame
}
