% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{add_tooltip}
\alias{add_tooltip}
\title{Add tooltips to nodes for visNetwork visualization.}
\usage{
add_tooltip(nodes_df)
}
\arguments{
\item{nodes_df}{Data frame of nodes with columns: KEGG, label, source, value.}
}
\value{
nodes_df with added 'title' column for tooltips.
}
\description{
Add tooltips to nodes for visNetwork visualization.
}
\details{
The tooltip includes a button to the specific KEGG entry page. If multiple KEGG IDs are present, they are concatenated with '+' in the URL. It also adds information about the node name, source of differential expression data, and value.
}
