% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping_interfaces.R,
%   R/mapping_functions.R
\name{annotation_dbi_interface}
\alias{annotation_dbi_interface}
\title{Map IDs using AnnotationDbi}
\usage{
annotation_dbi_interface(from, to, keys, organism = "human", db = NULL, ...)

annotation_dbi_interface(from, to, keys, organism = "human", db = NULL, ...)
}
\arguments{
\item{from}{Source ID type (e.g., "ENSEMBL", "SYMBOL", "   UNIPROT")}

\item{to}{Target ID type (e.g., "KEGG", "ENTREZID")}

\item{keys}{Vector of IDs to map}

\item{organism}{Organism name}
}
\value{
Named vector of mapped IDs, names are original IDs

Named vector of mapped IDs, names are original IDs
}
\description{
Map IDs using AnnotationDbi

Map IDs using AnnotationDbi
}
