% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8utils_find_relationships.R
\name{build_inchi_relationships}
\alias{build_inchi_relationships}
\title{Build inchi Relationships}
\usage{
build_inchi_relationships(
  rowdata_metabo,
  rowdata_target,
  get_ensembl = FALSE,
  get_uniprot = FALSE,
  progressbar = NULL
)
}
\arguments{
\item{rowdata_metabo}{A data frame or similar object containing metabolomics data.
This should include inchi IDs.}

\item{rowdata_target}{A data frame or similar object containing target data.
The row names should correspond to UniProt or Ensembl IDs.}

\item{get_ensembl}{Logical. If TRUE, Ensembl gene IDs will be retrieved and included
in the output. Default is TRUE.}

\item{get_uniprot}{Logical. If TRUE, UniProt protein IDs will be retrieved and included
in the output. Default is TRUE.}
}
\value{
A data frame containing the following columns:
- \code{UNIPROT}: UniProt protein IDs (if \code{get_uniprot} is TRUE).
- \code{ENSEMBL}: Ensembl gene IDs (if \code{get_ensembl} is TRUE).
- \code{inchikey}: inchi IDs from the metabolomics data.
The UniProt and Ensembl IDs are filtered to include only those present in
\code{rowdata_target}. inchi IDs are repeated to match the lengths of the corresponding
UniProt/Ensembl data.
}
\description{
This function generates a data frame of relationships between inchi IDs, UniProt IDs,
and Ensembl IDs based on metabolomics and target data. It queries UniProt to retrieve
protein and gene information associated with the provided inchi IDs and filters the
results based on their presence in the target data.
}
\details{
The function performs the following steps:
\enumerate{
\item Extracts inchi IDs from the metabolomics data (\code{rowdata_metabo}).
\item Filters out NA values from the inchi IDs.
\item Queries UniProt for each inchi ID to retrieve associated protein and gene information.
\item Filters the retrieved UniProt and Ensembl IDs to include only those present in
the row names of \code{rowdata_target}.
\item Constructs a data frame containing the relationships between inchi IDs, UniProt IDs,
and Ensembl IDs.
}
}
\examples{
# Example usage:
# relationships <- build_inchi_relatioships(rowdata_metabo, rowdata_target)
}
