% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8utils_find_relationships.R
\name{build_uniprot_to_ensembl}
\alias{build_uniprot_to_ensembl}
\title{Build UniProt to Ensembl Mapping}
\usage{
build_uniprot_to_ensembl(rowdata_prot, rowdata_trans, organism)
}
\arguments{
\item{rowdata_prot}{A data frame or matrix containing protein data, where row names are UniProt IDs.}

\item{rowdata_trans}{A data frame or matrix containing transcript data, where row names are Ensembl IDs.}

\item{organism}{A character string specifying the organism. Use "Hs" for human or "Mm" for mouse.}
}
\value{
A data frame containing the mapping of UniProt IDs to Ensembl IDs, filtered to include
only those Ensembl IDs present in \code{rowdata_trans}.
}
\description{
This function maps UniProt identifiers to Ensembl identifiers for a given organism.
It uses the appropriate annotation database based on the specified organism and filters
the results to include only those Ensembl IDs present in the provided transcript data.
}
\details{
The function uses the \code{AnnotationDbi::select} function to perform the mapping. It checks for
missing values in the mapping and issues a warning if some entries could not be mapped. If
all entries are missing, the function stops with an error.
}
\note{
Ensure that the \code{org.Hs.eg.db} or \code{org.Mm.eg.db} package is installed and loaded, depending
on the organism specified.
}
\examples{
\dontrun{
# Example usage:
rowdata_prot <- data.frame(matrix(ncol = 0, nrow = 5))
rownames(rowdata_prot) <- c("P12345", "Q67890", "A11111", "B22222", "C33333")
rowdata_trans <- data.frame(matrix(ncol = 0, nrow = 3))
rownames(rowdata_trans) <- c("ENSG000001", "ENSG000002", "ENSG000003")

result <- build_uniprot_to_ensembl(rowdata_prot, rowdata_trans, organism = "Hs")
print(result)
}

}
