% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{check_de_entry}
\alias{check_de_entry}
\title{Check the structure and validity of each entry in de_results}
\usage{
check_de_entry(de_entry, name)
}
\arguments{
\item{de_entry}{An entry from the de_results list}

\item{name}{The name of the entry in the de_results list (for error messages)}
}
\value{
Invisible TRUE if all checks pass, otherwise stops with an error
}
\description{
Check the structure and validity of each entry in de_results
}
