% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{download_kgml}
\alias{download_kgml}
\title{Download and cache KEGG KGML files.}
\usage{
download_kgml(pathway_id, bfc)
}
\arguments{
\item{pathway_id}{KEGG pathway ID (e.g., "hsa04110").}

\item{cache_dir}{Directory to store cached KGML files (default: "kgml_cache").}
}
\value{
Path to the cached KGML file.
}
\description{
Download and cache KEGG KGML files.
}
