% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping_interfaces.R,
%   R/mapping_functions.R
\name{entrez_to_kegg}
\alias{entrez_to_kegg}
\title{Entrez ID → KEGG ID}
\usage{
entrez_to_kegg(entrez_ids, organism = "human")

entrez_to_kegg(entrez_ids, organism = "human")
}
\arguments{
\item{entrez_ids}{Vector of Entrez gene IDs}

\item{organism}{Organism name}
}
\value{
Named vector of KEGG gene IDs, names are original Entrez IDs

Named vector of KEGG gene IDs, names are original Entrez IDs
}
\description{
Map a vector of Entrez gene IDs to KEGG gene IDs using UniProt as an intermediary

Map a vector of Entrez gene IDs to KEGG gene IDs using UniProt as an intermediary
}
