% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{expand_keggs}
\alias{expand_keggs}
\title{Handle multiple KEGG IDs in a single string separated by ";"}
\usage{
expand_keggs(kegg_df)
}
\arguments{
\item{kegg_string}{Character string of KEGG IDs separated by ";"}
}
\value{
Data frame with a single column 'KEGG' containing individual KEGG IDs
}
\description{
Handle multiple KEGG IDs in a single string separated by ";"
}
