% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_compounds}
\alias{get_compounds}
\title{Get KEGG compounds with caching.}
\usage{
get_compounds(bfc)
}
\arguments{
\item{bfc}{A BiocFileCache object for caching.}
}
\value{
A named character vector of KEGG compounds.
}
\description{
Get KEGG compounds with caching.
}
