% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{get_kgml}
\alias{get_kgml}
\title{Download KEGG KGML file for a given pathway ID, with caching.}
\usage{
get_kgml(url)
}
\arguments{
\item{pathway_id}{KEGG pathway ID (e.g., "hsa04110" or "04110").}
}
\value{
Content of the respose as text, or NULL if download failed.
}
\description{
Download KEGG KGML file for a given pathway ID, with caching.
}
