% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_id_mapping.R
\name{map_ids}
\alias{map_ids}
\title{Map IDs using UniProt ID mapping service}
\usage{
map_ids(from_db, to_db, ids)
}
\arguments{
\item{from_db}{Source database (e.g., "UniProtKB_AC-ID")}

\item{to_db}{Target database (e.g., "KEGG")}

\item{ids}{Vector of IDs to map}
}
\value{
Named vector of mapped IDs, names are original IDs
}
\description{
Map IDs using UniProt ID mapping service
}
