% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{map_results_to_nodes}
\alias{map_results_to_nodes}
\title{Map differential expression results to nodes}
\usage{
map_results_to_nodes(g, de_results, return_type = "visNetwork")
}
\arguments{
\item{g}{An igraph object representing the pathway.}

\item{de_results}{Named list of differential expression results.}

\item{return_type}{Output type: "igraph" or "visNetwork".}
}
\value{
An igraph or visNetwork object with mapped results.
}
\description{
Map differential expression results to nodes
}
