% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9main_app.R
\name{MovidaApp}
\alias{MovidaApp}
\title{Launch the MOVIDA Shiny Application}
\usage{
MovidaApp(movida_list)
}
\arguments{
\item{movida_list}{A list containing the data required to initialize the MOVIDA data model.}
}
\value{
The function starts the Shiny application and does not return any value.
}
\description{
This function initializes and launches the MOVIDA Shiny application using the provided list of MOVIDA data.
It loads required scripts, initializes the global data model, and starts the Shiny app.
}
\details{
The function sources UI, server, and data model scripts, initializes the global \code{movida_data} object,
loads relationships from the \code{relationships/} directory, and launches the Shiny application.
}
