% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgml_interface.R
\name{parse_kgml_entries}
\alias{parse_kgml_entries}
\title{Parse KEGG KGML files to extract nodes and edges data frames.}
\usage{
parse_kgml_entries(file)
}
\arguments{
\item{file}{Path to the KGML XML file.}
}
\value{
A tibble with columns: id, name, type, link, reaction, graphics_name, fgcolor, bgcolor, graphics_type, x, y, width, height.
}
\description{
Parse KEGG KGML files to extract nodes and edges data frames.
}
