% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgml_interface.R
\name{parse_kgml_relations}
\alias{parse_kgml_relations}
\title{Parse KEGG KGML files to extract relations and edges data frames.}
\usage{
parse_kgml_relations(file)
}
\arguments{
\item{file}{Path to the KGML XML file.}
}
\value{
A tibble with columns: from, to, type, name, value.
}
\description{
Parse KEGG KGML files to extract relations and edges data frames.
}
