% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_kegg_prefix}
\alias{remove_kegg_prefix}
\title{Convert KEGG IDs with prefixes to IDs without prefixes}
\usage{
remove_kegg_prefix(kegg_ids)
}
\arguments{
\item{kegg_ids}{Character vector of KEGG IDs with prefixes (e.g., "cpd:C00022", "mmu:1234", "ko:K00001 ko:K00002")}
}
\value{
List of character vectors with KEGG IDs without prefixes
}
\description{
Convert KEGG IDs with prefixes to IDs without prefixes
}
