% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping_interfaces.R,
%   R/mapping_functions.R
\name{select_database}
\alias{select_database}
\title{Select the appropriate AnnotationDbi database based on organism}
\usage{
select_database(organism)

select_database(organism)
}
\arguments{
\item{organism}{Organism name ("human" or "mouse")}
}
\value{
The corresponding AnnotationDbi database

The corresponding AnnotationDbi database

The corresponding AnnotationDbi database
}
\description{
Select the appropriate AnnotationDbi database based on organism

Select the appropriate AnnotationDbi database based on organism
 
 @param organism Organism name ("human" or "mouse")
}
