% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{style_edges}
\alias{style_edges}
\title{Style edges based on their subtype for visNetwork visualization.}
\usage{
style_edges(edges_df)
}
\arguments{
\item{edges_df}{Data frame of edges with a column 'subtype'.}
}
\value{
edges_df with added visual styling columns: color, dashes, arrows, label.
}
\description{
Style edges based on their subtype for visNetwork visualization.
}
