% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{style_nodes}
\alias{style_nodes}
\title{Style nodes based on their type for visNetwork visualization.}
\usage{
style_nodes(nodes_df, node_size_multiplier = 1.2)
}
\arguments{
\item{nodes_df}{Data frame of nodes with a column 'type'.}

\item{node_size_multiplier}{Numeric factor to scale node sizes (default: 1.2).}
}
\value{
nodes_df with added visual styling columns: shape, fixed, widthConstraint, heightConstraint, size.
}
\description{
Style nodes based on their type for visNetwork visualization.
}
