% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{to_organism_kegg}
\alias{to_organism_kegg}
\title{Convert organism name or abbreviation to KEGG organism code}
\usage{
to_organism_kegg(organism)
}
\arguments{
\item{organism}{Organism name or abbreviation (e.g., "human", "hs", "mouse", "mm")}
}
\value{
KEGG organism code (e.g., "hsa" for human, "mmu" for mouse)
}
\description{
Convert organism name or abbreviation to KEGG organism code
}
