% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{uniprot_to_kegg}
\alias{uniprot_to_kegg}
\title{UniProt ID → KEGG ID
Map a vector of UniProt IDs to KEGG gene IDs using UniProt ID mapping service}
\usage{
uniprot_to_kegg(...)
}
\arguments{
\item{uniprot_ids}{Vector of UniProt IDs}

\item{organism}{Organism name}
}
\value{
Named vector of KEGG gene IDs, names are original UniProt IDs
}
\description{
UniProt ID → KEGG ID
Map a vector of UniProt IDs to KEGG gene IDs using UniProt ID mapping service
}
