% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkegg_to_visnetwork.R
\name{map_results_to_nodes}
\alias{map_results_to_nodes}
\title{Map differential expression results to nodes}
\usage{
map_results_to_nodes(nodes_df, de_results)
}
\arguments{
\item{nodes_df}{Data frame of nodes from KGML.}

\item{de_results}{Named list of DE results.}
}
\value{
Nodes data frame with mapped results and colors.
}
\description{
Map differential expression results to nodes
}
